fix provides for eigen feature extractor and remove old code and clean up
